@@ -23,12 +23,11 @@ from sys import modules
 from typing import (
     TYPE_CHECKING,
     Any,
-    Callable,
     Literal,
     Optional,
     TypeVar,
-    Union,
     cast,
+    overload,
 )
 
 import numpy as np
@@ -37,7 +36,7 @@ import pytensor.sparse as sparse
 import pytensor.tensor as pt
 import scipy.sparse as sps
 
-from pytensor.compile import DeepCopyOp, get_mode
+from pytensor.compile import DeepCopyOp, Function, get_mode
 from pytensor.compile.sharedvalue import SharedVariable
 from pytensor.graph.basic import Constant, Variable, graph_inputs
 from pytensor.scalar import Cast
@@ -132,18 +131,18 @@ class ContextMeta(type):
 
     # FIXME: is there a more elegant way to automatically add methods to the class that
     # are instance methods instead of class methods?
-    def __init__(cls, name, bases, nmspc, context_class: Optional[type] = None, **kwargs):
+    def __init__(cls, name, bases, nmspc, context_class: type | None = None, **kwargs):
         """Add ``__enter__`` and ``__exit__`` methods to the new class automatically."""
         if context_class is not None:
             cls._context_class = context_class
         super().__init__(name, bases, nmspc)
 
-    def get_context(cls, error_if_none=True, allow_block_model_access=False) -> Optional[T]:
+    def get_context(cls, error_if_none=True, allow_block_model_access=False) -> T | None:
         """Return the most recently pushed context object of type ``cls``
         on the stack, or ``None``. If ``error_if_none`` is True (default),
         raise a ``TypeError`` instead of returning ``None``."""
         try:
-            candidate: Optional[T] = cls.get_contexts()[-1]
+            candidate: T | None = cls.get_contexts()[-1]
         except IndexError:
             # Calling code expects to get a TypeError if the entity
             # is unfound, and there's too much to fix.
@@ -184,7 +183,7 @@ class ContextMeta(type):
     # than a class.
     @property
     def context_class(cls) -> type:
-        def resolve_type(c: Union[type, str]) -> type:
+        def resolve_type(c: type | str) -> type:
             if isinstance(c, str):
                 c = getattr(modules[cls.__module__], c)
             if isinstance(c, type):
@@ -194,7 +193,7 @@ class ContextMeta(type):
         assert cls is not None
         if isinstance(cls._context_class, str):
             cls._context_class = resolve_type(cls._context_class)
-        if not isinstance(cls._context_class, (str, type)):
+        if not isinstance(cls._context_class, str | type):
             raise ValueError(
                 f"Context class for {cls.__name__}, {cls._context_class}, is not of the right type"
             )
@@ -615,7 +614,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def compile_logp(
         self,
-        vars: Optional[Union[Variable, Sequence[Variable]]] = None,
+        vars: Variable | Sequence[Variable] | None = None,
         jacobian: bool = True,
         sum: bool = True,
         **compile_kwargs,
@@ -637,7 +636,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def compile_dlogp(
         self,
-        vars: Optional[Union[Variable, Sequence[Variable]]] = None,
+        vars: Variable | Sequence[Variable] | None = None,
         jacobian: bool = True,
         **compile_kwargs,
     ) -> PointFunc:
@@ -655,7 +654,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def compile_d2logp(
         self,
-        vars: Optional[Union[Variable, Sequence[Variable]]] = None,
+        vars: Variable | Sequence[Variable] | None = None,
         jacobian: bool = True,
         **compile_kwargs,
     ) -> PointFunc:
@@ -673,10 +672,10 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def logp(
         self,
-        vars: Optional[Union[Variable, Sequence[Variable]]] = None,
+        vars: Variable | Sequence[Variable] | None = None,
         jacobian: bool = True,
         sum: bool = True,
-    ) -> Union[Variable, list[Variable]]:
+    ) -> Variable | list[Variable]:
         """Elemwise log-probability of the model.
 
         Parameters
@@ -697,7 +696,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         varlist: list[TensorVariable]
         if vars is None:
             varlist = self.free_RVs + self.observed_RVs + self.potentials
-        elif not isinstance(vars, (list, tuple)):
+        elif not isinstance(vars, list | tuple):
             varlist = [vars]
         else:
             varlist = cast(list[TensorVariable], vars)
@@ -752,7 +751,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def dlogp(
         self,
-        vars: Optional[Union[Variable, Sequence[Variable]]] = None,
+        vars: Variable | Sequence[Variable] | None = None,
         jacobian: bool = True,
     ) -> Variable:
         """Gradient of the models log-probability w.r.t. ``vars``.
@@ -772,7 +771,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         if vars is None:
             value_vars = None
         else:
-            if not isinstance(vars, (list, tuple)):
+            if not isinstance(vars, list | tuple):
                 vars = [vars]
 
             value_vars = []
@@ -791,7 +790,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def d2logp(
         self,
-        vars: Optional[Union[Variable, Sequence[Variable]]] = None,
+        vars: Variable | Sequence[Variable] | None = None,
         jacobian: bool = True,
     ) -> Variable:
         """Hessian of the models log-probability w.r.t. ``vars``.
@@ -811,7 +810,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         if vars is None:
             value_vars = None
         else:
-            if not isinstance(vars, (list, tuple)):
+            if not isinstance(vars, list | tuple):
                 vars = [vars]
 
             value_vars = []
@@ -926,7 +925,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         return self.free_RVs + self.deterministics
 
     @property
-    def coords(self) -> dict[str, Union[tuple, None]]:
+    def coords(self) -> dict[str, tuple | None]:
         """Coordinate values for model dimensions."""
         return self._coords
 
@@ -956,10 +955,10 @@ class Model(WithMemoization, metaclass=ContextMeta):
     def add_coord(
         self,
         name: str,
-        values: Optional[Sequence] = None,
-        mutable: Optional[bool] = None,
+        values: Sequence | None = None,
+        mutable: bool | None = None,
         *,
-        length: Optional[Union[int, Variable]] = None,
+        length: int | Variable | None = None,
     ):
         """Registers a dimension coordinate with the model.
 
@@ -1000,7 +999,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         if name in self.coords:
             if not np.array_equal(values, self.coords[name]):
                 raise ValueError(f"Duplicate and incompatible coordinate: {name}.")
-        if length is not None and not isinstance(length, (int, Variable)):
+        if length is not None and not isinstance(length, int | Variable):
             raise ValueError(
                 f"The `length` passed for the '{name}' coord must be an int, PyTensor Variable or None."
             )
@@ -1014,9 +1013,9 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def add_coords(
         self,
-        coords: dict[str, Optional[Sequence]],
+        coords: dict[str, Sequence | None],
         *,
-        lengths: Optional[dict[str, Optional[Union[int, Variable]]]] = None,
+        lengths: dict[str, int | Variable | None] | None = None,
     ):
         """Vectorized version of ``Model.add_coord``."""
         if coords is None:
@@ -1026,7 +1025,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         for name, values in coords.items():
             self.add_coord(name, values, length=lengths.get(name, None))
 
-    def set_dim(self, name: str, new_length: int, coord_values: Optional[Sequence] = None):
+    def set_dim(self, name: str, new_length: int, coord_values: Sequence | None = None):
         """Update a mutable dimension.
 
         Parameters
@@ -1072,7 +1071,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def set_initval(self, rv_var, initval):
         """Sets an initial value (strategy) for a random variable."""
-        if initval is not None and not isinstance(initval, (Variable, str)):
+        if initval is not None and not isinstance(initval, Variable | str):
             # Convert scalars or array-like inputs to ndarrays
             initval = rv_var.type.filter(initval)
 
@@ -1081,8 +1080,8 @@ class Model(WithMemoization, metaclass=ContextMeta):
     def set_data(
         self,
         name: str,
-        values: Union[Sequence, np.ndarray],
-        coords: Optional[dict[str, Sequence]] = None,
+        values: Sequence | np.ndarray,
+        coords: dict[str, Sequence] | None = None,
     ):
         """Changes the values of a data variable in the model.
 
@@ -1288,8 +1287,8 @@ class Model(WithMemoization, metaclass=ContextMeta):
         rv_var: TensorVariable,
         data: np.ndarray,
         dims,
-        transform: Union[Any, None],
-        total_size: Union[int, None],
+        transform: Any | None,
+        total_size: int | None,
     ) -> TensorVariable:
         """Create a `TensorVariable` for an observed random variable.
 
@@ -1371,7 +1370,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
         return rv_var
 
     def create_value_var(
-        self, rv_var: TensorVariable, transform: Any, value_var: Optional[Variable] = None
+        self, rv_var: TensorVariable, transform: Any, value_var: Variable | None = None
     ) -> TensorVariable:
         """Create a ``TensorVariable`` that will be used as the random
         variable's "value" in log-likelihood graphs.
@@ -1429,7 +1428,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
         return value_var
 
-    def add_named_variable(self, var, dims: Optional[tuple[Union[str, None], ...]] = None):
+    def add_named_variable(self, var, dims: tuple[str | None, ...] | None = None):
         """Add a random graph variable to the named variables of the model.
 
         This can include several types of variables such basic_RVs, Data, Deterministics,
@@ -1526,15 +1525,37 @@ class Model(WithMemoization, metaclass=ContextMeta):
             rvs_to_transforms=self.rvs_to_transforms,
         )
 
+    @overload
     def compile_fn(
         self,
-        outs: Union[Variable, Sequence[Variable]],
+        outs: Variable | Sequence[Variable],
         *,
-        inputs: Optional[Sequence[Variable]] = None,
+        inputs: Sequence[Variable] | None = None,
+        mode=None,
+        point_fn: Literal[True] = True,
+        **kwargs,
+    ) -> PointFunc: ...
+
+    @overload
+    def compile_fn(
+        self,
+        outs: Variable | Sequence[Variable],
+        *,
+        inputs: Sequence[Variable] | None = None,
+        mode=None,
+        point_fn: Literal[False],
+        **kwargs,
+    ) -> Function: ...
+
+    def compile_fn(
+        self,
+        outs: Variable | Sequence[Variable],
+        *,
+        inputs: Sequence[Variable] | None = None,
         mode=None,
         point_fn: bool = True,
         **kwargs,
-    ) -> Union[PointFunc, Callable[[Sequence[np.ndarray]], Sequence[np.ndarray]]]:
+    ) -> PointFunc | Function:
         """Compiles an PyTensor function
 
         Parameters
@@ -1724,7 +1745,7 @@ class Model(WithMemoization, metaclass=ContextMeta):
 
     def debug(
         self,
-        point: Optional[dict[str, np.ndarray]] = None,
+        point: dict[str, np.ndarray] | None = None,
         fn: Literal["logp", "dlogp", "random"] = "logp",
         verbose: bool = False,
     ):
@@ -1871,10 +1892,10 @@ class Model(WithMemoization, metaclass=ContextMeta):
     def to_graphviz(
         self,
         *,
-        var_names: Optional[Iterable[VarName]] = None,
+        var_names: Iterable[VarName] | None = None,
         formatting: str = "plain",
-        save: Optional[str] = None,
-        figsize: Optional[tuple[int, int]] = None,
+        save: str | None = None,
+        figsize: tuple[int, int] | None = None,
         dpi: int = 300,
     ):
         """Produce a graphviz Digraph from a PyMC model.
@@ -2039,14 +2060,14 @@ def set_data(new_data, model=None, *, coords=None):
 
 
 def compile_fn(
-    outs: Union[Variable, Sequence[Variable]],
+    outs: Variable | Sequence[Variable],
     *,
-    inputs: Optional[Sequence[Variable]] = None,
+    inputs: Sequence[Variable] | None = None,
     mode=None,
     point_fn: bool = True,
-    model: Optional[Model] = None,
+    model: Model | None = None,
     **kwargs,
-) -> Union[PointFunc, Callable[[Sequence[np.ndarray]], Sequence[np.ndarray]]]:
+) -> PointFunc | Function:
     """Compiles an PyTensor function
 
     Parameters
