@@ -12,7 +12,6 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from typing import Union
 
 from pytensor.compile import SharedVariable
 from pytensor.graph.basic import Constant, walk
@@ -49,7 +48,7 @@ def str_for_dist(
             dist_args = [
                 _str_for_input_var(x, formatting=formatting)
                 for x in dist.owner.inputs
-                if not isinstance(x, (RandomStateSharedVariable, RandomGeneratorSharedVariable))
+                if not isinstance(x, RandomStateSharedVariable | RandomGeneratorSharedVariable)
             ]
 
     print_name = dist.name
@@ -169,10 +168,10 @@ def _str_for_input_var(var: Variable, formatting: str) -> str:
             # in case other code overrides str_repr, fallback
             return False
 
-    if isinstance(var, (Constant, SharedVariable)):
+    if isinstance(var, Constant | SharedVariable):
         return _str_for_constant(var, formatting)
     elif isinstance(
-        var.owner.op, (RandomVariable, SymbolicRandomVariable)
+        var.owner.op, RandomVariable | SymbolicRandomVariable
     ) or _is_potential_or_deterministic(var):
         # show the names for RandomVariables, Deterministics, and Potentials, rather
         # than the full expression
@@ -195,7 +194,7 @@ def _str_for_input_rv(var: Variable, formatting: str) -> str:
         return _str
 
 
-def _str_for_constant(var: Union[Constant, SharedVariable], formatting: str) -> str:
+def _str_for_constant(var: Constant | SharedVariable, formatting: str) -> str:
     if isinstance(var, Constant):
         var_data = var.data
         var_type = "constant"
@@ -219,13 +218,13 @@ def _str_for_expression(var: Variable, formatting: str) -> str:
 
     # construct a string like f(a1, ..., aN) listing all random variables a as arguments
     def _expand(x):
-        if x.owner and (not isinstance(x.owner.op, (RandomVariable, SymbolicRandomVariable))):
+        if x.owner and (not isinstance(x.owner.op, RandomVariable | SymbolicRandomVariable)):
             return reversed(x.owner.inputs)
 
     parents = [
         x
         for x in walk(nodes=var.owner.inputs, expand=_expand)
-        if x.owner and isinstance(x.owner.op, (RandomVariable, SymbolicRandomVariable))
+        if x.owner and isinstance(x.owner.op, RandomVariable | SymbolicRandomVariable)
     ]
     names = [x.name for x in parents]
 
@@ -254,7 +253,7 @@ def _latex_escape(text: str) -> str:
     return text.replace("$", r"\$")
 
 
-def _default_repr_pretty(obj: Union[TensorVariable, Model], p, cycle):
+def _default_repr_pretty(obj: TensorVariable | Model, p, cycle):
     """Handy plug-in method to instruct IPython-like REPLs to use our str_repr above."""
     # we know that our str_repr does not recurse, so we can ignore cycle
     try:
