@@ -24,8 +24,7 @@ from collections.abc import Iterator, Mapping, Sequence
 from typing import (
     Any,
     Literal,
-    Optional,
-    Union,
+    TypeAlias,
     overload,
 )
 
@@ -38,7 +37,7 @@ from pytensor.graph.basic import Variable
 from rich.console import Console
 from rich.progress import Progress
 from rich.theme import Theme
-from typing_extensions import Protocol, TypeAlias
+from typing_extensions import Protocol
 
 import pymc as pm
 
@@ -81,7 +80,7 @@ __all__ = [
     "init_nuts",
 ]
 
-Step: TypeAlias = Union[BlockedStep, CompoundStep]
+Step: TypeAlias = BlockedStep | CompoundStep
 
 
 class SamplingIteratorCallback(Protocol):
@@ -98,8 +97,8 @@ def instantiate_steppers(
     model: Model,
     steps: list[Step],
     selected_steps: Mapping[type[BlockedStep], list[Any]],
-    step_kwargs: Optional[dict[str, dict]] = None,
-) -> Union[Step, list[Step]]:
+    step_kwargs: dict[str, dict] | None = None,
+) -> Step | list[Step]:
     """Instantiate steppers assigned to the model variables.
 
     This function is intended to be called automatically from ``sample()``, but
@@ -156,10 +155,10 @@ def instantiate_steppers(
 
 def assign_step_methods(
     model: Model,
-    step: Optional[Union[Step, Sequence[Step]]] = None,
-    methods: Optional[Sequence[type[BlockedStep]]] = None,
-    step_kwargs: Optional[dict[str, Any]] = None,
-) -> Union[Step, list[Step]]:
+    step: Step | Sequence[Step] | None = None,
+    methods: Sequence[type[BlockedStep]] | None = None,
+    step_kwargs: dict[str, Any] | None = None,
+) -> Step | list[Step]:
     """Assign model variables to appropriate step methods.
 
     Passing a specified model will auto-assign its constituent stochastic
@@ -193,7 +192,7 @@ def assign_step_methods(
     assigned_vars: set[Variable] = set()
 
     if step is not None:
-        if isinstance(step, (BlockedStep, CompoundStep)):
+        if isinstance(step, BlockedStep | CompoundStep):
             steps.append(step)
         else:
             steps.extend(step)
@@ -264,13 +263,13 @@ def _sample_external_nuts(
     tune: int,
     chains: int,
     target_accept: float,
-    random_seed: Union[RandomState, None],
-    initvals: Union[StartDict, Sequence[Optional[StartDict]], None],
+    random_seed: RandomState | None,
+    initvals: StartDict | Sequence[StartDict | None] | None,
     model: Model,
-    var_names: Optional[Sequence[str]],
+    var_names: Sequence[str] | None,
     progressbar: bool,
-    idata_kwargs: Optional[dict],
-    nuts_sampler_kwargs: Optional[dict],
+    idata_kwargs: dict | None,
+    nuts_sampler_kwargs: dict | None,
     **kwargs,
 ):
     if nuts_sampler_kwargs is None:
@@ -376,25 +375,25 @@ def sample(
     draws: int = 1000,
     *,
     tune: int = 1000,
-    chains: Optional[int] = None,
-    cores: Optional[int] = None,
+    chains: int | None = None,
+    cores: int | None = None,
     random_seed: RandomState = None,
     progressbar: bool = True,
-    progressbar_theme: Optional[Theme] = default_progress_theme,
+    progressbar_theme: Theme | None = default_progress_theme,
     step=None,
-    var_names: Optional[Sequence[str]] = None,
+    var_names: Sequence[str] | None = None,
     nuts_sampler: Literal["pymc", "nutpie", "numpyro", "blackjax"] = "pymc",
-    initvals: Optional[Union[StartDict, Sequence[Optional[StartDict]]]] = None,
+    initvals: StartDict | Sequence[StartDict | None] | None = None,
     init: str = "auto",
     jitter_max_retries: int = 10,
     n_init: int = 200_000,
-    trace: Optional[TraceOrBackend] = None,
+    trace: TraceOrBackend | None = None,
     discard_tuned_samples: bool = True,
     compute_convergence_checks: bool = True,
     keep_warning_stat: bool = False,
     return_inferencedata: Literal[True] = True,
-    idata_kwargs: Optional[dict[str, Any]] = None,
-    nuts_sampler_kwargs: Optional[dict[str, Any]] = None,
+    idata_kwargs: dict[str, Any] | None = None,
+    nuts_sampler_kwargs: dict[str, Any] | None = None,
     callback=None,
     mp_ctx=None,
     **kwargs,
@@ -406,28 +405,28 @@ def sample(
     draws: int = 1000,
     *,
     tune: int = 1000,
-    chains: Optional[int] = None,
-    cores: Optional[int] = None,
+    chains: int | None = None,
+    cores: int | None = None,
     random_seed: RandomState = None,
     progressbar: bool = True,
-    progressbar_theme: Optional[Theme] = default_progress_theme,
+    progressbar_theme: Theme | None = default_progress_theme,
     step=None,
-    var_names: Optional[Sequence[str]] = None,
+    var_names: Sequence[str] | None = None,
     nuts_sampler: Literal["pymc", "nutpie", "numpyro", "blackjax"] = "pymc",
-    initvals: Optional[Union[StartDict, Sequence[Optional[StartDict]]]] = None,
+    initvals: StartDict | Sequence[StartDict | None] | None = None,
     init: str = "auto",
     jitter_max_retries: int = 10,
     n_init: int = 200_000,
-    trace: Optional[TraceOrBackend] = None,
+    trace: TraceOrBackend | None = None,
     discard_tuned_samples: bool = True,
     compute_convergence_checks: bool = True,
     keep_warning_stat: bool = False,
     return_inferencedata: Literal[False],
-    idata_kwargs: Optional[dict[str, Any]] = None,
-    nuts_sampler_kwargs: Optional[dict[str, Any]] = None,
+    idata_kwargs: dict[str, Any] | None = None,
+    nuts_sampler_kwargs: dict[str, Any] | None = None,
     callback=None,
     mp_ctx=None,
-    model: Optional[Model] = None,
+    model: Model | None = None,
     **kwargs,
 ) -> MultiTrace: ...
 
@@ -436,30 +435,30 @@ def sample(
     draws: int = 1000,
     *,
     tune: int = 1000,
-    chains: Optional[int] = None,
-    cores: Optional[int] = None,
+    chains: int | None = None,
+    cores: int | None = None,
     random_seed: RandomState = None,
     progressbar: bool = True,
-    progressbar_theme: Optional[Theme] = default_progress_theme,
+    progressbar_theme: Theme | None = default_progress_theme,
     step=None,
-    var_names: Optional[Sequence[str]] = None,
+    var_names: Sequence[str] | None = None,
     nuts_sampler: Literal["pymc", "nutpie", "numpyro", "blackjax"] = "pymc",
-    initvals: Optional[Union[StartDict, Sequence[Optional[StartDict]]]] = None,
+    initvals: StartDict | Sequence[StartDict | None] | None = None,
     init: str = "auto",
     jitter_max_retries: int = 10,
     n_init: int = 200_000,
-    trace: Optional[TraceOrBackend] = None,
+    trace: TraceOrBackend | None = None,
     discard_tuned_samples: bool = True,
     compute_convergence_checks: bool = True,
     keep_warning_stat: bool = False,
     return_inferencedata: bool = True,
-    idata_kwargs: Optional[dict[str, Any]] = None,
-    nuts_sampler_kwargs: Optional[dict[str, Any]] = None,
+    idata_kwargs: dict[str, Any] | None = None,
+    nuts_sampler_kwargs: dict[str, Any] | None = None,
     callback=None,
     mp_ctx=None,
-    model: Optional[Model] = None,
+    model: Model | None = None,
     **kwargs,
-) -> Union[InferenceData, MultiTrace]:
+) -> InferenceData | MultiTrace:
     r"""Draw samples from the posterior using the given step methods.
 
     Multiple step methods are supported via compound step methods.
@@ -846,7 +845,7 @@ def sample(
 
 def _sample_return(
     *,
-    run: Optional[RunType],
+    run: RunType | None,
     traces: Sequence[IBaseTrace],
     tune: int,
     t_sampling: float,
@@ -856,7 +855,7 @@ def _sample_return(
     keep_warning_stat: bool,
     idata_kwargs: dict[str, Any],
     model: Model,
-) -> Union[InferenceData, MultiTrace]:
+) -> InferenceData | MultiTrace:
     """Final step of `pm.sampler` that picks/slices chains,
     runs diagnostics and converts to the desired return type."""
     # Pick and slice chains to keep the maximum number of samples
@@ -945,9 +944,9 @@ def _sample_many(
     chains: int,
     traces: Sequence[IBaseTrace],
     start: Sequence[PointType],
-    random_seed: Optional[Sequence[RandomSeed]],
+    random_seed: Sequence[RandomSeed] | None,
     step: Step,
-    callback: Optional[SamplingIteratorCallback] = None,
+    callback: SamplingIteratorCallback | None = None,
     **kwargs,
 ):
     """Samples all chains sequentially.
@@ -989,8 +988,8 @@ def _sample(
     step: Step,
     trace: IBaseTrace,
     tune: int,
-    model: Optional[Model] = None,
-    progressbar_theme: Optional[Theme] = default_progress_theme,
+    model: Model | None = None,
+    progressbar_theme: Theme | None = default_progress_theme,
     callback=None,
     **kwargs,
 ) -> None:
@@ -1056,9 +1055,9 @@ def _iter_sample(
     trace: IBaseTrace,
     chain: int = 0,
     tune: int = 0,
-    model: Optional[Model] = None,
+    model: Model | None = None,
     random_seed: RandomSeed = None,
-    callback: Optional[SamplingIteratorCallback] = None,
+    callback: SamplingIteratorCallback | None = None,
 ) -> Iterator[bool]:
     """Generator for sampling one chain. (Used in singleprocess sampling.)
 
@@ -1138,10 +1137,10 @@ def _mp_sample(
     random_seed: Sequence[RandomSeed],
     start: Sequence[PointType],
     progressbar: bool = True,
-    progressbar_theme: Optional[Theme] = default_progress_theme,
+    progressbar_theme: Theme | None = default_progress_theme,
     traces: Sequence[IBaseTrace],
-    model: Optional[Model] = None,
-    callback: Optional[SamplingIteratorCallback] = None,
+    model: Model | None = None,
+    callback: SamplingIteratorCallback | None = None,
     mp_ctx=None,
     **kwargs,
 ) -> None:
@@ -1222,8 +1221,8 @@ def _mp_sample(
 
 def _init_jitter(
     model: Model,
-    initvals: Optional[Union[StartDict, Sequence[Optional[StartDict]]]],
-    seeds: Union[Sequence[int], np.ndarray],
+    initvals: StartDict | Sequence[StartDict | None] | None,
+    seeds: Sequence[int] | np.ndarray,
     jitter: bool,
     jitter_max_retries: int,
 ) -> list[PointType]:
@@ -1279,12 +1278,12 @@ def init_nuts(
     init: str = "auto",
     chains: int = 1,
     n_init: int = 500_000,
-    model: Optional[Model] = None,
+    model: Model | None = None,
     random_seed: RandomSeed = None,
     progressbar=True,
     jitter_max_retries: int = 10,
-    tune: Optional[int] = None,
-    initvals: Optional[Union[StartDict, Sequence[Optional[StartDict]]]] = None,
+    tune: int | None = None,
+    initvals: StartDict | Sequence[StartDict | None] | None = None,
     **kwargs,
 ) -> tuple[Sequence[PointType], NUTS]:
     """Set up the mass matrix initialization for NUTS.
