@@ -89,7 +89,7 @@ class Censored(Distribution):
     @classmethod
     def dist(cls, dist, lower, upper, **kwargs):
         if not isinstance(dist, TensorVariable) or not isinstance(
-            dist.owner.op, (RandomVariable, SymbolicRandomVariable)
+            dist.owner.op, RandomVariable | SymbolicRandomVariable
         ):
             raise ValueError(
                 f"Censoring dist must be a distribution created via the `.dist()` API, got {type(dist)}"
