@@ -11,6 +11,7 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
+import itertools
 import warnings
 
 import numpy as np
@@ -178,7 +179,7 @@ class Mixture(Distribution):
 
     @classmethod
     def dist(cls, w, comp_dists, **kwargs):
-        if not isinstance(comp_dists, (tuple, list)):
+        if not isinstance(comp_dists, tuple | list):
             # comp_dists is a single component
             comp_dists = [comp_dists]
         elif len(comp_dists) == 1:
@@ -204,7 +205,7 @@ class Mixture(Distribution):
             # TODO: Allow these to not be a RandomVariable as long as we can call `ndim_supp` on them
             #  and resize them
             if not isinstance(dist, TensorVariable) or not isinstance(
-                dist.owner.op, (RandomVariable, SymbolicRandomVariable)
+                dist.owner.op, RandomVariable | SymbolicRandomVariable
             ):
                 raise ValueError(
                     f"Component dist must be a distribution created via the `.dist()` API, got {type(dist)}"
@@ -480,7 +481,7 @@ def marginal_mixture_default_transform(op, rv):
                 transform.backward(value, *component.owner.inputs)
                 for transform, component in zip(default_transforms, components)
             ]
-            for expr1, expr2 in zip(backward_expressions[:-1], backward_expressions[1:]):
+            for expr1, expr2 in itertools.pairwise(backward_expressions):
                 if not equal_computations([expr1], [expr2]):
                     transform_warning()
                     return None
