@@ -26,7 +26,6 @@ from pymc.step_methods import (
     Slice,
 )
 from pymc.step_methods.compound import (
-    BlockedStep,
     StatsBijection,
     flatten_steps,
     get_stats_dtypes_shapes_from_steps,
@@ -38,10 +37,7 @@ from tests.models import simple_2model_continuous
 
 
 def test_all_stepmethods_emit_tune_stat():
-    attrs = [getattr(pm.step_methods, n) for n in dir(pm.step_methods)]
-    step_types = [
-        attr for attr in attrs if isinstance(attr, type) and issubclass(attr, BlockedStep)
-    ]
+    step_types = pm.step_methods.STEP_METHODS
     assert len(step_types) > 5
     for cls in step_types:
         assert "tune" in cls.stats_dtypes_shapes
