@@ -152,7 +152,7 @@ def test_extract_obs_data():
     constant = pt.as_tensor(data_m.filled())
     z_at = pt.set_subtensor(constant[mask.nonzero()], missing_values)
 
-    assert isinstance(z_at.owner.op, (AdvancedIncSubtensor, AdvancedIncSubtensor1))
+    assert isinstance(z_at.owner.op, AdvancedIncSubtensor | AdvancedIncSubtensor1)
 
     res = extract_obs_data(z_at)
 
@@ -169,7 +169,7 @@ def test_extract_obs_data():
     constant = pt.as_tensor(data_m.filled())
     z_at = pt.set_subtensor(constant[mask.nonzero()], missing_values)
 
-    assert isinstance(z_at.owner.op, (AdvancedIncSubtensor, AdvancedIncSubtensor1))
+    assert isinstance(z_at.owner.op, AdvancedIncSubtensor | AdvancedIncSubtensor1)
 
     res = extract_obs_data(z_at)
 
